requirements.txt
torch
diffusers
transformers
accelerate
opencv-python
Pillow
